Add Bluebird config
The goal here is to disable the warning about a promise being created but not returned. This happens because of Tarn (used by Knex) but its not really an issue and it can be safelly ignored (check tarn's github issue #5 for more info). I've added all the config keys, just in case I need it.